% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yproject.R
\name{as_proj_spec}
\alias{as_proj_spec}
\title{Create a project object from data spec objects}
\usage{
as_proj_spec(..., output = tempfile(fileext = ".yml"))
}
\arguments{
\item{...}{yspec objects}

\item{output}{the name and path where the project file is to be written}
}
\value{
an object of class yproj
}
\description{
Create a project object from data spec objects
}
