% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_data.R
\name{axis_labs}
\alias{axis_labs}
\alias{axis_col_labs}
\alias{axis_label}
\title{Generate axis labels or col/label structures for plots}
\usage{
axis_labs(.spec, ..., .fun = axis_label)

axis_col_labs(.spec, ..., .fun = axis_label)

axis_label(x)
}
\arguments{
\item{.spec}{a yspec object}

\item{...}{unquoted column names in the spec}

\item{.fun}{a function that forms the axis label data for}

\item{x}{a ycol object
a single column}
}
\description{
Generate axis labels or col/label structures for plots
}
