% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spec.R
\name{check_data}
\alias{check_data}
\alias{check_data_file}
\title{Check a data set against its specification}
\usage{
check_data(data, spec)

check_data_file(data, file)
}
\arguments{
\item{data}{a data frame}

\item{spec}{a yspec object}

\item{file}{the full path to a yaml specification file}
}
\description{
Check a data set against its specification
}
