% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definetemplate.R
\name{definetemplate}
\alias{definetemplate}
\alias{mrgtemplate}
\title{Create build location for templated define documents}
\usage{
definetemplate(root = tempdir())

mrgtemplate(...)
}
\arguments{
\item{root}{directory where the build will take place}

\item{...}{arguments passed to [definetemplate()] when [mrgtemplate()] is 
called}
}
\value{
The path to the build directory.
}
\description{
Create build location for templated define documents
}
\details{
The following files will be copied into the
build directory: \verb{_output.yml}, \code{definetemplate.tex},
\code{header.tex}.
}
\examples{

ans <- definetemplate()

ans

list.files(ans)
}
