% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{df_to_yspec}
\alias{df_to_yspec}
\title{Convert a data frame to a yspec yaml file}
\usage{
df_to_yspec(
  x,
  desc = "",
  projectid = "",
  varname,
  vardesc = NULL,
  varunit = NULL,
  vartype = NULL,
  output = "yspec.yaml",
  varcodelist = NULL,
  codelist_func = parse_yspec_codelist
)
}
\arguments{
\item{x}{data frame to convert}

\item{desc}{description parameter for the yspec file}

\item{projectid}{projectid parameter for the yspec file}

\item{varname}{column specifying variable names to include in the yspec file}

\item{vardesc}{column specifying variable descriptions to include in the yspec file}

\item{varunit}{column specifying variable units to include in the yspec file}

\item{vartype}{column specifying variable types to include in the yspec file}

\item{output}{name of the output yspec file}

\item{varcodelist}{column containing coded values (comma- or newline-separated)}

\item{codelist_func}{function taking a codelist string and returning a named list of codes}
}
\description{
Convert a data frame to a yspec yaml file
}
\author{
Omar I. Elashkar
}
