% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exspec.R
\name{ex_spec}
\alias{ex_spec}
\title{Generate an example specification object}
\usage{
ex_spec(file = "spec.yml")
}
\arguments{
\item{file}{an example spec file name}
}
\description{
Generate an example specification object
}
