% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{get_meta}
\alias{get_meta}
\title{Get meta data from a specification object}
\usage{
get_meta(x)
}
\arguments{
\item{x}{a yspec object}
}
\description{
Get meta data from a specification object
}
\examples{
spec <- ys_help$spec()

ans <- get_meta(spec)

}
