% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataspec.R
\name{load_spec}
\alias{load_spec}
\title{Load a data specification file}
\usage{
load_spec(file)
}
\arguments{
\item{file}{name of yaml file containing specification}
}
\description{
Load a data specification file
}
