% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spec.R
\name{load_spec_any}
\alias{load_spec_any}
\title{Load a specification file, guessing the type}
\usage{
load_spec_any(file, ...)
}
\arguments{
\item{file}{a yaml file name}

\item{...}{arguments passed to \link{ys_load} or \link{ys_load_proj}}
}
\description{
Load a specification file, guessing the type
}
\examples{

class(load_spec_any(file_spec_ex()))

class(load_spec_any(file_proj_ex()))

}
