% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yproject.R
\name{load_spec_proj}
\alias{load_spec_proj}
\alias{ys_load_proj}
\title{Load a project-wide spec file}
\usage{
load_spec_proj(file)

ys_load_proj(...)
}
\arguments{
\item{file}{file name}
}
\description{
Load a project-wide spec file
}
\examples{

file <- file_proj_ex()

spec <- load_spec_proj(file)

spec
}
