% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-lscape.R
\name{lscape_table}
\alias{lscape_table}
\title{Create a data frame for landscape define document}
\usage{
lscape_table(spec, title_case = FALSE)
}
\arguments{
\item{spec}{a `yspec` object}

\item{title_case}{if `TRUE` the `Label` column is passed to 
[tools::toTitleCase()]}
}
\value{
A tibble with columns:
- `Variable`
- `Label`
- `Unit`
- `Type`
- `Decode/Source/Comment`
}
\description{
Create a data frame for landscape define document
}
