% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_outline.R
\name{md_outline}
\alias{md_outline}
\title{Generate an outline representation of a data spec}
\usage{
md_outline(x)
}
\arguments{
\item{x}{a yspec object}
}
\value{
character
}
\description{
Generate an outline representation of a data spec
}
