% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgtemmplate.R
\name{mrgtemplate}
\alias{mrgtemplate}
\title{Create build location for branded documents}
\usage{
mrgtemplate(root = tempdir())
}
\arguments{
\item{root}{directory where the build will take place}
}
\value{
The path to the build directory.
}
\description{
The idea is to create a build directory such that, when 
you render
}
\details{
The following files will be copied into the 
build directory: \code{_output.yml}, \code{mrgtemplate.tex}, 
\code{header.tex}, \code{mrglogo.pdf}.
}
\examples{

ans <- mrgtemplate()

ans

list.files(ans)

}
