% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_input.R
\name{nm_input}
\alias{nm_input}
\title{Create NONMEM $INPUT block code from yspec object}
\usage{
nm_input(
  spec,
  .width = 65,
  .cat = TRUE,
  .long = FALSE,
  .drop = NULL,
  .decodes = FALSE,
  ...
)
}
\arguments{
\item{spec}{a yspec object.}

\item{.width}{passed to \code{\link[base:strwrap]{base::strwrap()}} to limit the output line length.}

\item{.cat}{if \code{TRUE}, the text is sent to the console with \code{\link[=cat]{cat()}}.}

\item{.long}{if \code{TRUE}, produce \verb{$INPUT} in long, verbose format.}

\item{.drop}{a character vector or comma-separated string of columns to
drop in the \verb{$INPUT} listing; columns with character type are automatically
dropped, so there is no need to list them here.}

\item{.decodes}{if \code{TRUE}, print \code{value} and \code{decode} information where
available for discrete column data; this is only printed when \code{.long = TRUE}.}

\item{...}{unquoted column rename pairs with format
\verb{<new name> = <old name>}.}
}
\value{
A character vector of text forming the \verb{$INPUT} block (including
the \verb{$INPUT} part).
}
\description{
Extract column names from the yspec object and format for inclusion in
a NONMEM control stream, potentially renaming or dropping columns.
The default output is in wide format, including only the column names as
well as any rename or drop information. The long format puts one column
on a different line and includes the short name and optionally column
decode information. See examples.
}
\details{
Columns with character type are automatically dropped; there is no need
to list these under the \code{.drop} argument.
}
\examples{
spec <- ys_help$spec()
nm_input(spec)
nm_input(spec, DOSE = AMT, .drop = "ALT,BMI")
nm_input(spec, .long = TRUE)
nm_input(spec, .long = TRUE, .decodes = TRUE)

}
