% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_table.R
\name{pander_table}
\alias{pander_table}
\title{Render a spec as a pandoc table}
\usage{
pander_table(x, data_file = "data", ...)
}
\arguments{
\item{x}{a spec object}

\item{data_file}{not used}

\item{...}{passed to other functions}
}
\description{
Render a spec as a pandoc table
}
