% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{primary_keys}
\alias{primary_keys}
\title{Get the primary keys from a specification object}
\usage{
primary_keys(x)
}
\arguments{
\item{x}{a yspec object}
}
\description{
Get the primary keys from a specification object
}
