% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{print_fda_define}
\alias{print_fda_define}
\alias{fda_define}
\title{Print a define document suitable for FDA submission}
\usage{
print_fda_define(file, main = "Datasets")

fda_define(...)
}
\arguments{
\item{file}{full path to define yaml file}

\item{main}{Title for the main heading}

\item{...}{passed to \code{\link{print_fda_define}}}
}
\description{
Print a define document suitable for FDA submission
}
\details{
Use \code{print_fda_define} to generate the output in the context
of an Rmarkdown document.  Use \code{fda_define} to generate the
document as a character vector.
}
