% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{pull_meta}
\alias{pull_meta}
\title{Pull a single item from the meta data object}
\usage{
pull_meta(x, what)
}
\arguments{
\item{x}{a yspec object}

\item{what}{character name of item in meta}
}
\description{
Pull a single item from the meta data object
}
\examples{

spec <- ys_help$spec()

pull_meta(spec, "description")

}
