% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{short}
\alias{short}
\alias{short.ycol}
\alias{short.yspec}
\title{Form a short name}
\usage{
short(x, ...)

\method{short}{ycol}(
  x,
  default = "col",
  short_max = getOption("yspec.short.max", NULL),
  title_case = getOption("yspec.short.title", FALSE),
  ...
)

\method{short}{yspec}(x, default = "short", .aslist = TRUE, ...)
}
\arguments{
\item{x}{a yspec or ycol object}

\item{...}{passed to methods}

\item{default}{character field name to use when the \code{short} field is not
found}

\item{short_max}{if \code{short} is found, but contains more than \code{short_max}
characters, \code{col} will be used}

\item{title_case}{\code{short} will be converted with \link[tools:toTitleCase]{tools::toTitleCase} if it
is found}

\item{.aslist}{if `TRUE`a named list is returned; otherwise, a named character
vector}
}
\description{
Form a short name
}
