% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aaaa.R
\name{specification}
\alias{specification}
\alias{SETUP}
\title{Yaml specification reference}
\description{
Yaml specification reference
}
\section{SETUP__}{

Include this block at the top of the yaml file; this is where you
enter meta information about the data set.
\itemize{
\item \code{description}: \code{<character>}; a short descrption of the data set
\item \code{projectnumber}: \code{<character>} the project reference number; may be
incorporated into rendered define documents
\item \code{lookup_file}: \code{<character>}; a yaml array of other yaml files where 
yspec will look for column lookup inforomation
\item \code{use_internal_db}: \code{<logical> (true/false)}; if \code{true}, then yspec will
load the internal column lookup database
\item \code{glue}: \code{<map>}; specify name/value pairs;  in the yaml data specification,
use \code{<<name>>} in the text and \code{value} will glued into the text after
it has been sanitized; intended use is to allow LaTex code to evade the
sanitizer
}
}

\section{Column specification syntax}{
NA
}

