% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_spec_list}
\alias{test_spec_list}
\title{Test specification code}
\usage{
test_spec_list(x)
}
\arguments{
\item{x}{a list column specification data}
}
\description{
For internal / testing use
}
\keyword{internal}
