% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{update_short}
\alias{update_short}
\title{Update short names in a yspec object}
\usage{
update_short(spec, ...)
}
\arguments{
\item{spec}{a yspec object}

\item{...}{`<column name> = <short name>` pairs}
}
\description{
Update short names in a yspec object
}
\details{
If no update items are passed in, the original spec object will be returned.
An error will be issued if a column update is requested, but can't be found
in the spec.
}
\examples{
sp <- ys_help$spec()

sp2 <- update_short(sp, ID = "subject", ALB = "serum albumin")

}
