% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_table.R
\name{x_table}
\alias{x_table}
\title{Format specification output with xtable}
\usage{
x_table(x, ...)
}
\arguments{
\item{x}{a \code{yspec} object}

\item{...}{not used}
}
\description{
Requires the xtable package to be installed.  \link{pander_table_df} actually
formats the table.
}
\examples{
spec <- load_spec_ex()

x_table(spec)

}
