% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-note.R
\name{ys_col_note}
\alias{ys_col_note}
\title{Create table note text for abbreviated columns}
\usage{
ys_col_note(
  .spec,
  ...,
  .unit = FALSE,
  .sep = "; ",
  .to_string = TRUE,
  .width = NULL
)
}
\arguments{
\item{.spec}{a yspec object.}

\item{...}{passed to \code{\link[=ys_select]{ys_select()}}.}

\item{.unit}{passed to \code{\link[=ys_get_short]{ys_get_short()}} or \code{\link[=ys_get_short_unit]{ys_get_short_unit()}}.}

\item{.sep}{a separator character for columns.}

\item{.to_string}{logical; if \code{TRUE}, then a single string is returned.}

\item{.width}{if \code{numeric} and \code{.to_string} is \code{TRUE}, then the result is
passed through \code{\link[base:strwrap]{base::strwrap()}}.}
}
\value{
A string of length one when \code{.to_string} is \code{TRUE} or a character vector
if \code{.to_string} is \code{FALSE}.
}
\description{
Create table note text for abbreviated columns
}
\examples{

spec <- ys_help$spec()

ys_col_note(spec, AST, ALT, SCR, .unit = TRUE) 

}
