% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{data}
\name{ys_demo}
\alias{ys_demo}
\title{Demonstration object}
\format{An object of class \code{environment} of length 5.}
\usage{
ys_demo
}
\description{
Demonstration object
}
\details{
\itemize{
\item \code{file} is the file name and path of a demonstration yspec file
\item \code{spec()} is a function that generates a yspec object from \code{file}
\item \code{data()} is a function that generates a data set that can be checked
against \code{spec()}
\item \code{csv} is the file name (and path) to the data set generated by \code{data()}
\item \code{proj()} is a function that generates a project object that can be
rendered into a document
}
}
\examples{
spec <- ys_demo$spec()

ys_demo$file

dat <- ys_demo$data()  

}
\keyword{datasets}
