% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{ys_extend}
\alias{ys_extend}
\title{Extend a yspec object}
\usage{
ys_extend(x, file)
}
\arguments{
\item{x}{A \code{yspec} object.}

\item{file}{The path to a yaml specification file to load and join to \code{x}.}
}
\description{
Use this function to read and join another spec file to an existing
\code{yspec} object. The additional spec file can contain additional columns
that might be added to the data set on an ad-hoc basis or could include
modeling outputs (e.g. \code{IPRED}).
}
\details{
The extension \code{yspec} object cannot contain any columns in common with the
primary \code{yspec} object.
}
\examples{
extension_file <- system.file("spec", "nm-extension.yml")

spec <- yys_help$spec() 
spec <- ys_extend(spec, extension_file)
tail(spec)

}
