% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-tidy.R
\name{ys_fill_dots}
\alias{ys_fill_dots}
\title{Supplement data in the dots list with default values}
\usage{
ys_fill_dots(x, ..., .overwrite = FALSE)
}
\arguments{
\item{...}{\code{name = value} data to be added to \code{dots}}

\item{.overwrite}{logical indicating whether \code{x} should be overwritten with
data in \code{...}}
}
\description{
Please avoid thinking of this function as a way to mutate \code{dots}.  It sort of
is but not in the way you think. It really is for supplementing the spec
so that downstream functions are guaranteed to find these names with
sensible values for every column in the spec. mutate will be a different
function.
}
\details{
By default, this function will not overwrite data in \code{dots} when it already
exists. This behavior can be changed with the \code{.overwrite} argument. Note
that when \code{.overwrite} is \code{TRUE}, all columns in the cpec will have the
same value in \code{dots} for the \code{name} getting set.
}
\examples{
spec <- ys_help$spec()

spec2 <- ys_fill_dots(spec, new_var = FALSE)

}
