% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{ys_get_assets}
\alias{ys_get_assets}
\title{View example workflow assets}
\usage{
ys_get_assets(output, overwrite = FALSE)
}
\arguments{
\item{output}{the directory where assets should be copied}

\item{overwrite}{passed to \link{file.copy}}
}
\description{
This function copies files out of the yspec install directory and into
the directory of your choosing.
}
\details{
\itemize{
\item \code{analysis1.yml}: an example yaml file
\item \code{analysis1.csv}: data set that works with \code{analysis1.yml}
\item \code{ysdb_internal.yml}: the internal column lookup data base
\item \code{ys_example.Rmd}: contains testing code that utilizes other assets
\item \code{ys_get_assets.md}: a listing of assets that is exported
}
}
\examples{
ys_get_assets("yspec_stuff")


}
