% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_short}
\alias{ys_get_short}
\title{Get short}
\usage{
ys_get_short(x, short_max = Inf, title_case = FALSE, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{short_max}{if \code{short} is found, but contains more than \code{short_max}
characters, \code{col} will be used}

\item{title_case}{\code{short} will be converted with \link[tools:toTitleCase]{tools::toTitleCase} if it
is found}

\item{...}{passed to short methods; see details}
}
\description{
Forms a short for a column (\code{ycol} method) or columns (\code{yspec} method). Use
\link[purrr:map]{purrr::map} or \link[purrr:map_chr]{purrr::map_chr} to simply extract the short field.
}
\examples{
spec <- ys_help$spec()

ys_get_short(spec)

ys_get_short(spec$WT)

\dontrun{
purrr:::map(spec,"short")
}
}
\seealso{
\link{ys_get_unit}, \link{ys_get_label}, \link{ys_get_short_unit}
}
