% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_short_unit}
\alias{ys_get_short_unit}
\title{Get short with unit}
\usage{
ys_get_short_unit(x, .aslist = TRUE, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{.aslist}{a named list is returned if \code{TRUE}, otherwise a named
character vector}

\item{...}{arguments passed to \link{ys_get_short} and \link{ys_get_unit}}
}
\description{
Get short with unit
}
\examples{
spec <- ys_help$spec()

ys_get_short_unit(spec)

ys_get_short_unit(spec$WT)

}
\seealso{
\link{ys_get_short}, \link{ys_get_unit}, \link{ys_get_label}
}
