% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-tidy.R
\name{ys_join}
\alias{ys_join}
\title{Join yspec objects together}
\usage{
ys_join(left, right, ...)
}
\arguments{
\item{left}{a yspec object.}

\item{right}{a yspec object.}

\item{...}{more yspec objects.}
}
\value{
A single yspec object with \code{left}, \code{right} and \code{...} all joined together.
}
\description{
Join yspec objects together
}
\details{
All inputs must be \code{yspec} objects. When the \code{right} spec (or specs passed
under \code{...}) is joined to the \code{left} spec, columns that exist in both
\code{left} and \code{right} are removed from \code{right} before joining.
}
\examples{

spec <- ys_help$spec()

l <- ys_select(spec, WT, BMI)
r <- ys_select(spec, TIME, TAD)
rr <- ys_select(spec, EVID, MDV, CMT, BMI)

ys_join(l, r)

ys_join(l, r, rr)

}
\seealso{
\code{\link[=ys_rename]{ys_rename()}}, \code{\link[=ys_filter]{ys_filter()}}, \code{\link[=ys_select]{ys_select()}}
}
