% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spec.R
\name{ys_load}
\alias{ys_load}
\alias{ys_load_file}
\alias{load_spec}
\title{Load a data specification file}
\usage{
ys_load(file, verbose = FALSE, ...)

ys_load_file(file, data_path = NULL, data_stem = NULL, verbose = FALSE, ...)

load_spec(...)
}
\arguments{
\item{file}{Name of yaml file containing specification.}

\item{verbose}{Logical: print information to the console as the file
is processed.}

\item{...}{Other arguments to update \code{SETUP__}.}

\item{data_path}{Optional path to data sets.}

\item{data_stem}{Optional alternate stem for data files.}
}
\description{
Load a data specification file
}
\examples{

sp <- ys_load(ys_help$file())
sp

sp <- ys_load(ys_help$file(), verbose = TRUE)

}
