% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yproject.R
\name{ys_load_proj}
\alias{ys_load_proj}
\alias{load_spec_proj}
\title{Load a project-wide spec file}
\usage{
ys_load_proj(file)

load_spec_proj(...)
}
\arguments{
\item{file}{file name}

\item{...}{used for alias only}
}
\description{
Load a project-wide spec file
}
\examples{

file <- file_proj_ex()

spec <- ys_load_proj(file)

spec
}
