% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{ys_namespace}
\alias{ys_namespace}
\title{Change namespace}
\usage{
ys_namespace(x, namespace = NULL)
}
\arguments{
\item{x}{a \code{yspec} object.}

\item{namespace}{namespace name (character) to switch to.}
}
\value{
the \code{yspec} object (\code{x}) is returned, possibly modified.
}
\description{
Change namespace
}
