% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yproject.R
\name{ys_project}
\alias{ys_project}
\alias{ys_project_file}
\alias{as_proj_spec}
\title{Create a project object from data spec objects}
\usage{
ys_project(
  ...,
  output = tempfile(fileext = ".yml"),
  data_path = NULL,
  dots = list(),
  sponsor = "[sponsor]",
  projectnumber = "[projectnumber]",
  keep_spec_data = FALSE
)

ys_project_file(
  ...,
  output = tempfile(fileext = ".yml"),
  where = NULL,
  dots = list()
)

as_proj_spec(...)
}
\arguments{
\item{...}{yspec objects or file names of yaml specification files.}

\item{output}{the name and path where the project file is to be written.}

\item{data_path}{optional data path.}

\item{dots}{used to update \code{SETUP__} block data items.}

\item{sponsor}{optional project sponsor.}

\item{projectnumber}{optional project number in \code{...}.}

\item{keep_spec_data}{if \code{TRUE}, then the specification data is saved in the
meta space of the project object.}

\item{where}{directory containing the specification files if listed.}
}
\value{
an object of class yproj
}
\description{
Create a project object from data spec objects
}
\details{
Note that \code{sponsor} and \code{projectnumber} can be updated upon creating the
project object.  Other document-specific attributes (like \strong{title},
\strong{author}, \strong{date}) can be specified through \link{ys_document}.
}
\examples{
where <- system.file("spec", package = "yspec")

files <- file.path(where, c("DEM104101F_PK.yml", "DEM104101F_AE.yml"))

proj <- ys_project(ys_load(files[1]),ys_load(files[2]), sponsor="CompanyX")

proj <- ys_project_file(files[1], files[2], dots = list(projectnumber = "ZZZzzz"))

proj

}
