% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-tidy.R
\name{ys_recode}
\alias{ys_recode}
\title{Recode column names in a vector}
\usage{
ys_recode(col, spec, ..., unit = FALSE)
}
\arguments{
\item{col}{A character vector of names in \code{spec}.}

\item{spec}{A \code{yspec} object.}

\item{...}{Passed to \code{\link[=ys_get_short]{ys_get_short()}} (when \code{unit} is \code{FALSE}) or
\code{\link[=ys_get_short_unit]{ys_get_short_unit()}} (when \code{unit} is \code{TRUE}).}

\item{unit}{Logical indicating if the unit should be appended to the short
name.}
}
\description{
Use a \code{yspec} object to convert a vector of column names to the short
name with the unit optionally appended.
}
\examples{
spec <- ys_help$spec()

x <- c("WT", "BAR", "SCR", "TIME", "FOO")

ys_recode(x, spec)

ys_recode(x, ys_select(spec, -TIME), unit = TRUE, title_case = TRUE)

}
