% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-tidy.R
\name{ys_rename}
\alias{ys_rename}
\title{Rename spec columns}
\usage{
ys_rename(.x, ...)
}
\arguments{
\item{.x}{a yspec object}

\item{...}{tidy rename specification; use \code{new_name} = \code{old_name} to rename}
}
\value{
A yspec object
}
\description{
Rename spec columns
}
\examples{
spec <- ys_help$spec()

ans <- ys_rename(spec, TIME = TAFD, RENAL = RF)

tail(ans)

}
\seealso{
\code{\link[=ys_join]{ys_join()}}, \code{\link[=ys_filter]{ys_filter()}}, \code{\link[=ys_select]{ys_select()}}
}
