% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{ys_render}
\alias{ys_render}
\title{Render an arbitrary spec of project}
\usage{
ys_render(...)
}
\arguments{
\item{...}{passed to \link{render_define} or \link{render_fda_define}; it is important
to review these help topics to see what other aspects of the document
can be specified; see also \code{details} here.}
}
\description{
Render an arbitrary spec of project
}
