% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{ys_select_fl}
\alias{ys_select_fl}
\title{Select items from a yspec object via flags}
\usage{
ys_select_fl(x, ...)
}
\arguments{
\item{x}{a yspec object.}

\item{...}{tidy-select specification of flag names to select.}
}
\value{
A yspec object containing columns named in passed through \code{...}.
}
\description{
Select items from a yspec object via flags
}
\examples{

spec <- ys_help$spec()

ys_select_fl(spec, nm, times)

}
\seealso{
\code{\link[=ys_flags]{ys_flags()}}, \code{\link[=ys_flags_chr]{ys_flags_chr()}}
}
