% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{ys_spec_file}
\alias{ys_spec_file}
\title{Get the file name for a yspec object}
\usage{
ys_spec_file(x)
}
\arguments{
\item{x}{a yspec object}
}
\description{
This read from the yspec object meta data via [get_meta].
}
\examples{
spec <- load_spec_ex()
ys_spec_file(spec)

}
