% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{ys_view}
\alias{ys_view}
\alias{ys_view_db}
\title{View example specification file}
\usage{
ys_view(file = "analysis1.yml")

ys_view_db()
}
\arguments{
\item{file}{the example specification file to view}
}
\description{
Open up either an example data specification file or the `yspec`
internal data base file.
}
\details{
The specification file will be copied out of the yspec install 
location and into a temporary file.  That file will be opened as a regular
file in case `rstudioapi` is installed.  Otherwise, the file is opened
for viewing using [file.show].
}
\examples{
\dontrun{
ys_view()
ys_view_db()
}

}
