% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yspec_select_chr}
\alias{yspec_select_chr}
\title{Select all columns that are character}
\usage{
yspec_select_chr(.spec)
}
\arguments{
\item{.spec}{a yspec object}
}
\description{
Select all columns that are character
}
