% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yspec_select_discrete}
\alias{yspec_select_discrete}
\title{Select all columns that are discrete}
\usage{
yspec_select_discrete(.spec)
}
\arguments{
\item{.spec}{a yspec object}
}
\description{
Select all columns that are discrete
}
