% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{yspec_yml_file}
\alias{yspec_yml_file}
\title{Get the yaml file location for a spec object}
\usage{
yspec_yml_file(x, ...)
}
\arguments{
\item{x}{yspec object}

\item{...}{not used}
}
\description{
Get the yaml file location for a spec object
}
