% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yproject.R
\name{as_proj_spec}
\alias{as_proj_spec}
\alias{ys_project_file}
\alias{ys_project}
\title{Create a project object from data spec objects}
\usage{
as_proj_spec(..., output = tempfile(fileext = ".yml"),
  data_path = NULL, dots = list(), sponsor = "[sponsor]",
  projectnumber = "[projectnumber]")

ys_project_file(..., output = tempfile(fileext = ".yml"), where = NULL,
  dots = list())

ys_project(...)
}
\arguments{
\item{...}{yspec objects or file names of yaml specification files}

\item{output}{the name and path where the project file is to be written}

\item{data_path}{optional data path}

\item{dots}{used to update `SETUP__` block data items}

\item{sponsor}{optional project sponsor}

\item{projectnumber}{optional project number
in \code{...}}

\item{where}{directory containing the specification files if listed}
}
\value{
an object of class yproj
}
\description{
Create a project object from data spec objects
}
