% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{c.yspec}
\alias{c.yspec}
\title{Add extra column elements to a yspec object}
\usage{
\method{c}{yspec}(x, y, ..., .meta = get_meta(x))
}
\arguments{
\item{x}{a \code{yspec} object}

\item{y}{a \code{yspec} object}

\item{...}{not used}

\item{.meta}{meta information to use for the new spec object}
}
\description{
Add extra column elements to a yspec object
}
