% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{c.yspec}
\alias{c.yspec}
\title{Add extra column elements to a yspec object}
\usage{
\method{c}{yspec}(x, y, ...)
}
\arguments{
\item{x}{a `yspec` object}

\item{y}{a `yspec` object}
}
\description{
Add extra column elements to a yspec object
}
