% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_for_rmd}
\alias{define_for_rmd}
\title{Generate code for a generic define document}
\usage{
define_for_rmd(x, form_, proj = NULL, meta = NULL, tex = TRUE, ns = "tex")
}
\arguments{
\item{x}{a project file name}

\item{form_}{a function or the name of a function to format the spec
contents}

\item{proj}{a project object from which to render}

\item{meta}{meta data list}

\item{tex}{logical; if \code{TRUE} then try to switch to \code{tex} namespace if
it exists.}

\item{ns}{a character vector of namespaces to invoke prior to generating the
document.}
}
\description{
This function is for internal use by \code{\link[=render_define]{render_define()}}.
}
\keyword{internal}
