% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_for_rmd}
\alias{define_for_rmd}
\title{Generate code for a generic define document}
\usage{
define_for_rmd(yamlfile, format)
}
\arguments{
\item{yamlfile}{a project file name}

\item{format}{a function or the name of a function to format the spec
contents}
}
\description{
Generate code for a generic define document
}
\keyword{internal}
