% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{fda_content_table}
\alias{fda_content_table}
\alias{fda_content_table_file}
\title{Print a table of contents for FDA define document}
\usage{
fda_content_table(x)

fda_content_table_file(file)
}
\arguments{
\item{x}{a spec define object}

\item{file}{the full path to a yaml specification file}
}
\description{
Print a table of contents for FDA define document
}
\examples{
proj <- spec_ex_file("project.yml")
spec <- load_spec_proj(proj)
fda_content_table(spec)

}
\seealso{
\code{\link{load_spec_proj}}
}
