% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{fda_content_table}
\alias{fda_content_table}
\title{Print a table of contents for FDA define document}
\usage{
fda_content_table(x, ext = ".xpt", loc = ".")
}
\arguments{
\item{x}{a spec define object}

\item{ext}{data set file extension to include; this should only
be changed from default value of ".xpt" for testing purposes}

\item{loc}{location to use for data set files; this should
only be changed from default value of "." for testing purposes}
}
\description{
Print a table of contents for FDA define document
}
\examples{
proj <- file_proj_ex()

spec <- load_spec_proj(proj)

fda_content_table(spec)
}
\seealso{
\link{load_spec_proj}
}
