% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{fda_define}
\alias{fda_define}
\title{Generate content for FDA define document}
\usage{
fda_define(file, title = "Datasets")
}
\arguments{
\item{file}{full path to define yaml file}

\item{title}{used in yaml front matter}
}
\value{
A character vector of in markdown format.  Wrap
\code{fda_define} in \code{\link{writeLines}} and
render \code{asis} in an Rmarkdown document.
}
\description{
Generate content for FDA define document
}
\examples{
proj <- file_proj_ex()

cat(readLines(proj), sep = "\\n")

fda_define(proj)

}
