% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{fda_table}
\alias{fda_table}
\alias{fda_table_file}
\title{Generate a table for FDA define.pdf document}
\usage{
fda_table(x, widths = c(0.75, 2.1, 0.6, 2.2), ...)

fda_table_file(file, ns = c("tex", "define"))
}
\arguments{
\item{x}{a yspec object}

\item{widths}{column widths in inches; must be numeric vector with length 4}

\item{...}{not currently used}

\item{file}{the full path to yaml specification file}
}
\value{
Character vector of latex code for the content of an FDA
\code{define.pdf} document.  It includes a table of contents as well as data spec
tables for each dataset for a project.
}
\description{
Generate a table for FDA define.pdf document
}
\examples{
spec <- load_spec_ex()
spec
fda_table(spec)

}
