% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{get_values}
\alias{get_values}
\alias{get_values.yspec}
\alias{get_values.ycol}
\title{Extract vector of values from discrete data columns}
\usage{
get_values(x, ...)

\method{get_values}{yspec}(x, col, ...)

\method{get_values}{ycol}(x, drop = NULL, named = TRUE, ...)
}
\arguments{
\item{x}{a yspec object.}

\item{col}{the name of a column to extract.}

\item{drop}{values to drop; can be integer positions or decode labels.}

\item{named}{if `TRUE`, return will be named with decode labels if they
exist.}
}
\description{
Extract vector of values from discrete data columns
}
