% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exspec.R
\name{load_spec_ex}
\alias{load_spec_ex}
\title{Generate an example specification object}
\usage{
load_spec_ex(file = "spec.yml")
}
\arguments{
\item{file}{an example spec file name}
}
\description{
Generate an example specification object
}
