% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exspec.R
\name{load_spec_ex}
\alias{load_spec_ex}
\alias{load_proj_ex}
\alias{file_spec_ex}
\alias{file_proj_ex}
\title{Generate an example specification object}
\usage{
load_spec_ex(file = "spec.yml")

load_proj_ex(file = "project.yml")

file_spec_ex(file = "spec.yml")

file_proj_ex(file = "project.yml")
}
\arguments{
\item{file}{an example spec file name}
}
\description{
Generate an example specification object
}
\examples{

load_spec_ex()
load_proj_ex()
file_spec_ex()
file_proj_ex()

}
