% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mrgtemplate}
\alias{mrgtemplate}
\title{Build a directory with template requirements}
\usage{
mrgtemplate(root = tempdir())
}
\arguments{
\item{root}{directory where the build will take place}
}
\value{
The path to the build directory.
}
\description{
Build a directory with template requirements
}
\details{
The following files will be copied into the 
build directory: \code{_output.yml}, \code{mrgtemplate.tex}, 
\code{header.tex}, \code{mrglogo.pdf}.
}
