% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_input.R
\name{nm_input}
\alias{nm_input}
\title{Write columns and column information for NONMEM}
\usage{
nm_input(spec, width = 60)
}
\arguments{
\item{spec}{a yspec object}

\item{width}{number of}
}
\description{
Write columns and column information for NONMEM
}
