% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_table.R
\name{pander_table}
\alias{pander_table}
\alias{pander_table_df}
\title{Render a spec as a pandoc table}
\usage{
pander_table(x, ...)

pander_table_df(x)
}
\arguments{
\item{x}{a spec object}

\item{...}{passed to other functions}
}
\description{
Requires the pander package to be installed.  \link{pander_table_df} actually
formats the table.
}
\examples{
spec <- load_spec_ex()

pander_table_df(spec)

pander_table(spec)

}
