% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_table.R
\name{pander_table}
\alias{pander_table}
\alias{pander_table_df}
\title{DEPRECATED: Render a spec as a pandoc table}
\usage{
pander_table(x, ...)

pander_table_df(x, long = FALSE, ...)
}
\arguments{
\item{x}{a spec object}

\item{...}{passed to other functions}

\item{long}{if \code{TRUE}, then the table is made in a longer format}
}
\description{
Requires the pander package to be installed.  \link{pander_table_df} actually
formats the table.
}
\examples{
spec <- load_spec_ex()

\dontrun{
pander_table_df(spec)

pander_table(spec)
}

}
