% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_define}
\alias{render_define}
\alias{render_define.yproj}
\alias{render_define.character}
\alias{render_define.yspec}
\alias{render_spec}
\alias{render_spec.character}
\alias{render_spec.yspec}
\title{Render a \code{define.pdf} document}
\usage{
render_define(x, ...)

\method{render_define}{yproj}(
  x,
  stem = "define_working",
  format = "x_table_2",
  output_format = "pdf_document",
  output_dir = getwd(),
  build_dir = definetemplate(),
  title = "Data Specification",
  author = "",
  toc = "yes",
  number_sections = "yes",
  rmd_template = NULL,
  date = as.character(Sys.Date()),
  sponsor = NULL,
  projectnumber = NULL,
  ns = "tex",
  ...
)

\method{render_define}{character}(x, ..., dots = list())

\method{render_define}{yspec}(...)

render_spec(x, ...)

\method{render_spec}{character}(x, stem = basename(x), ..., dots = list())

\method{render_spec}{yspec}(x, stem = get_meta(x)[["name"]], ..., dots = list())
}
\arguments{
\item{x}{a \code{yproj} object or project specification file name.}

\item{...}{passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{stem}{used to name the output file.}

\item{format}{the name of a function that will generate code formatting
the data specification information.}

\item{output_format}{passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{output_dir}{passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{build_dir}{directory where \code{rmarkdown} should build the
document.}

\item{title}{used in yaml front matter.}

\item{author}{used in yaml front matter.}

\item{toc}{used in yaml front matter.}

\item{number_sections}{used in yaml front matter.}

\item{rmd_template}{full path to rmarkdown file to be used to template the
data specification document.}

\item{date}{used in yaml front matter.}

\item{sponsor}{optional project sponsor.}

\item{projectnumber}{optional project number in \code{...}.}

\item{ns}{character vector of namespaces to invoke prior to rendering the
document.}

\item{dots}{passed to object converter.}
}
\description{
\code{render_spec} is an alias to \code{render_define}.  See details.
}
\details{
\code{stem} should not include a file extension, just
the file stem.
}
\section{latex requirements}{


For all document types, the following \code{latex} packages are required:
\enumerate{
\item \code{array}
\item \code{longtable}
\item \code{booktabs}
\item \code{fontenc}
}

Make sure these packages are installed and available when trying to render a document.
}

\examples{

file <- ys_help$file()

file

\dontrun{
  render_define(file)
}
}
