% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_define}
\alias{render_define}
\title{Render a define.pdf document}
\usage{
render_define(file, stem = basename(file), format = c("pander_table",
  "md_outline"), output_format = "html_document", output_dir = getwd(),
  build_dir = tempdir(), title = "Data Specification",
  author = "MetrumRG", date = format(Sys.time()), ...)
}
\arguments{
\item{file}{a project spec file loaded via \code{\link{load_spec_proj}}}

\item{stem}{used to name the output file}

\item{format}{the name of a function that will genrate code formatting
the data specification information}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{build_dir}{directory where \code{rmarkdown} should build the
document}

\item{title}{used in yaml front matter}

\item{author}{used in yaml front matter}

\item{date}{used in yaml front matter}

\item{...}{passed to \code{rmarkdown::render}}
}
\description{
Render a define.pdf document
}
\details{
\code{stem} should not include a file extension, just
the file stem.
}
\examples{

file <- file_proj_ex()

file

render_define(file)

}
