% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_define}
\alias{render_define}
\alias{render_define.yproj}
\alias{render_define.character}
\alias{render_spec}
\alias{render_spec.character}
\alias{render_spec.yspec}
\title{Render a define.pdf document}
\usage{
render_define(x, ...)

\method{render_define}{yproj}(x, stem = "define", format = c("x_table",
  "pander_table", "md_outline"), output_format = "pdf_document",
  output_dir = getwd(), build_dir = tempdir(),
  title = "Data Specification", author = "MetrumRG", toc = "yes",
  number_sections = "yes", date = format(Sys.time()), ...)

\method{render_define}{character}(x, ...)

render_spec(x, ...)

\method{render_spec}{character}(x, stem = basename(x), ...)

\method{render_spec}{yspec}(x, stem = get_meta(x)[["name"]], ...)
}
\arguments{
\item{x}{a \code{yproj} object or project specification file name}

\item{...}{passed to \code{rmarkdown::render}}

\item{stem}{used to name the output file}

\item{format}{the name of a function that will genrate code formatting
the data specification information}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{build_dir}{directory where \code{rmarkdown} should build the
document}

\item{title}{used in yaml front matter}

\item{author}{used in yaml front matter}

\item{toc}{used in yaml front matter}

\item{number_sections}{used in yaml front matter}

\item{date}{used in yaml front matter}
}
\description{
`render_spec` generates a define document from a single specification object
or file name.
}
\details{
\code{stem} should not include a file extension, just
the file stem.
}
\examples{

file <- file_proj_ex()

file

\dontrun{
  render_define(file)
}
}
