% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_define}
\alias{render_define}
\title{Render a define.pdf document}
\usage{
render_define(file, title, output, ...)
}
\arguments{
\item{file}{yaml file identifying specifications}

\item{title}{a title for the define document}

\item{output}{character stem to create a name for the output file}

\item{...}{passed to \code{rmarkdown::render}}
}
\description{
Render a define.pdf document
}
\details{
\code{output} should not include a file extension.
}
