% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_define}
\alias{render_define}
\title{Render a define.pdf document}
\usage{
render_define(file, stem = basename(file), output = "define",
  output_format = "html_document", output_dir = getwd(),
  build_dir = tempdir(), title = "Data Specification",
  format = c("pander_table", "md_outline"), author = "MetrumRG",
  date = format(Sys.time()), ...)
}
\arguments{
\item{file}{a project spec file loaded via \code{\link{load_spec_proj}}}

\item{output}{character stem to create a name for the output file}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{title}{used for yaml front matter}

\item{author}{used for yaml front matter}

\item{date}{used for yaml front matter}

\item{...}{passed to \code{rmarkdown::render}}
}
\description{
Render a define.pdf document
}
\details{
\code{output} should not include a file extension, just
the file stem.
}
