% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_define}
\alias{render_define}
\title{Render a define.pdf document}
\usage{
render_define(file, output = "define", output_format = "html_document",
  output_dir = getwd(), title = "Data Specification", ...)
}
\arguments{
\item{file}{a project spec file loaded via \code{\link{load_spec_proj}}}

\item{output}{character stem to create a name for the output file}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{title}{a title for the define document}

\item{...}{passed to \code{rmarkdown::render}}
}
\description{
Render a define.pdf document
}
\details{
\code{output} should not include a file extension, just
the file stem.
}
