% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_define}
\alias{render_define}
\alias{render_define.yproj}
\alias{render_define.character}
\alias{render_define.yspec}
\alias{render_spec}
\alias{render_spec.character}
\alias{render_spec.yspec}
\title{Render a \code{define.pdf} document}
\usage{
render_define(x, ...)

\method{render_define}{yproj}(x, stem = "define_working",
  format = c("x_table", "pander_table", "md_outline"),
  output_format = "pdf_document", output_dir = getwd(),
  build_dir = tempdir(), title = "Data Specification",
  author = "MetrumRG", toc = "yes", number_sections = "yes",
  date = format(Sys.time()), ...)

\method{render_define}{character}(x, ..., dots = list())

\method{render_define}{yspec}(...)

render_spec(x, ...)

\method{render_spec}{character}(x, stem = basename(x), ...,
  dots = list())

\method{render_spec}{yspec}(x, stem = get_meta(x)[["name"]], ...,
  dots = list())
}
\arguments{
\item{x}{a \code{yproj} object or project specification file name}

\item{...}{passed to \link[rmarkdown:render]{rmarkdown::render}}

\item{stem}{used to name the output file}

\item{format}{the name of a function that will genrate code formatting
the data specification information}

\item{output_format}{passed to \link[rmarkdown:render]{rmarkdown::render}}

\item{output_dir}{passed to \link[rmarkdown:render]{rmarkdown::render}}

\item{build_dir}{directory where \code{rmarkdown} should build the
document}

\item{title}{used in yaml front matter}

\item{author}{used in yaml front matter}

\item{toc}{used in yaml front matter}

\item{number_sections}{used in yaml front matter}

\item{date}{used in yaml front matter}

\item{dots}{passed to object converter}
}
\description{
\code{render_spec} is an alias to \code{render_define}.  See detils.
}
\details{
\code{stem} should not include a file extension, just
the file stem.
}
\examples{

file <- ys_help$file()

file

\dontrun{
  render_define(file)
}
}
