% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{render_fda_define}
\alias{render_fda_define}
\alias{render_fda_define.yproj}
\alias{render_fda_define.character}
\title{Render a define document for sending to FDA}
\usage{
render_fda_define(x, ...)

\method{render_fda_define}{yproj}(x, ...)

\method{render_fda_define}{character}(x, stem = "define",
  title = "Data Definitions", date = format(Sys.time()),
  author = "MetrumRG Staff Scientist", output_dir = getwd(),
  build_dir = tempdir(), ...)
}
\arguments{
\item{x}{a yaml specification file name or a yproj object}

\item{...}{passed to \code{rmarkdown::render}}

\item{stem}{used to name the output document}

\item{title}{a title for the document}

\item{date}{the document date}

\item{author}{the document author}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{build_dir}{directory where rmarkdown will build the document}
}
\description{
Render a define document for sending to FDA
}
\examples{
proj_file <- file_proj_ex()

proj_file

\dontrun{
  render_fda_define(proj_file)
}

}
