% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{render_fda_define}
\alias{render_fda_define}
\alias{render_fda_define.yproj}
\alias{render_fda_define.character}
\alias{render_fda_define.yspec}
\title{Render a define document conforming to FDA standards}
\usage{
render_fda_define(x, ...)

\method{render_fda_define}{yproj}(x, stem = "define",
  title = "Data Definitions", date = format(Sys.time()),
  author = "MetrumRG Staff Scientist", format = "fda_define",
  output_dir = getwd(), build_dir = tempdir(), ext = ".xpt",
  loc = ".", ...)

\method{render_fda_define}{character}(x, ..., dots = list())

\method{render_fda_define}{yspec}(x, ..., dots = list())
}
\arguments{
\item{x}{a yaml specification file name or a yproj object}

\item{...}{passed to \code{rmarkdown::render}}

\item{stem}{used to name the output document}

\item{title}{a title for the document}

\item{date}{the document date}

\item{author}{the document author}

\item{format}{function to generate the define text}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{build_dir}{directory where rmarkdown will build the document}

\item{ext}{data set file extension to include; this should only 
be changed from default value of ".xpt" for testing puroses}

\item{loc}{location to use for data set files; this should 
only be changed from default value of "." for testing purposes}

\item{dots}{named list of arguments passed to object converter function}
}
\description{
Render a define document conforming to FDA standards
}
\examples{
proj_file <- file_proj_ex()

proj_file

\dontrun{
  render_fda_define(proj_file)
}

}
