% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_spec}
\alias{render_spec}
\title{Render a data specification object}
\usage{
render_spec(x, stem, format = c("md_outline"),
  output_format = "pdf_document", ...)
}
\arguments{
\item{x}{object}

\item{stem}{for output file name}

\item{format}{function defining how to render the object}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{...}{passed to \code{rmarkdown::render}}
}
\description{
Render a data specification object
}
