% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_spec}
\alias{render_spec}
\title{Render a data specification object}
\usage{
render_spec(yamlfile, stem = basename(yamlfile), data_file = data_stem(x),
  format = c("pander_table", "md_outline"), title = "Data Specification",
  author = "MetrumRG", template = NULL, date = format(Sys.time()),
  output_format = "html_document", output_dir = getwd(),
  build_dir = tempdir(), ...)
}
\arguments{
\item{stem}{for output file name}

\item{data_file}{the data file the spec is describing}

\item{format}{function defining how to render the object}

\item{title}{used for yaml front matter}

\item{author}{used for yaml front matter}

\item{date}{used for yaml front matter}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{build_dir}{where to build the document}

\item{...}{passed to \code{rmarkdown::render}}

\item{x}{object}
}
\description{
Render a data specification object
}
