% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_spec}
\alias{render_spec}
\title{Render a data specification object}
\usage{
render_spec(x, stem, data_file = data_stem(x), format = c("pander_table",
  "md_outline"), title = "Data Specification",
  output_format = "html_document", output_dir = getwd(),
  build_dir = tempdir(), ...)
}
\arguments{
\item{x}{object}

\item{stem}{for output file name}

\item{data_file}{the data file the spec is describing}

\item{format}{function defining how to render the object}

\item{title}{a title for the rendered document}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{build_dir}{where to build the document}

\item{...}{passed to \code{rmarkdown::render}}
}
\description{
Render a data specification object
}
