% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{render_spec}
\alias{render_spec}
\alias{render_spec.character}
\alias{render_spec.yspec}
\title{Render a data specification file}
\usage{
render_spec(x, ...)

\method{render_spec}{character}(x, stem = basename(x), format = c("x_table",
  "pander_table", "md_outline"), title = "Data Specification",
  author = "MetrumRG", date = format(Sys.time()),
  output_format = "pdf_document", output_dir = getwd(),
  build_dir = tempdir(), ...)

\method{render_spec}{yspec}(x, ...)
}
\arguments{
\item{x}{a data specification file name or a yspec object}

\item{...}{passed to \code{rmarkdown::render}}

\item{stem}{for output file name}

\item{format}{function defining how to render the object}

\item{title}{used for yaml front matter}

\item{author}{used for yaml front matter}

\item{date}{used for yaml front matter}

\item{output_format}{passed to \code{rmarkdown::render}}

\item{output_dir}{passed to \code{rmarkdown::render}}

\item{build_dir}{where to build the document}
}
\description{
Render a data specification file
}
\examples{

file <- file_spec_ex()
render_spec(file)

spec <- load_spec_ex()
render_spec(spec)

}
