% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_factor.R
\name{ys_add_factors}
\alias{ys_add_factors}
\alias{yspec_add_factors}
\alias{ys_make_factor}
\alias{yspec_make_factor}
\title{Add factors to data set based on spec information}
\usage{
ys_add_factors(
  .data,
  .spec,
  ...,
  .all = TRUE,
  .suffix = getOption("ys.fct.suffix", "_f")
)

yspec_add_factors(
  .data,
  .spec,
  ...,
  .all = TRUE,
  .suffix = getOption("ys.fct.suffix", "_f")
)

ys_make_factor(values, x, strict = TRUE)

yspec_make_factor(values, x, strict = TRUE)
}
\arguments{
\item{.data}{the data set to modify}

\item{.spec}{a yspec object}

\item{...}{unquoted column names for modification}

\item{.all}{if \code{TRUE} then any column with a \code{values} attribute or  where
the \code{make_factor} field evaluates to \code{TRUE} will be added as a factor}

\item{.suffix}{used to make the column name for the factors}

\item{values}{a vector of values to convert to a factor}

\item{x}{a ycol object}

\item{strict}{if `FALSE`, then an factor will be returned for any `values` type}
}
\description{
Add factors to data set based on spec information
}
\details{
Note that \code{.suffix} can be chosen using option \code{ys.fct.suffix}.
}
\examples{

spec <- load_spec_ex()

ys_make_factor(c(1,0,1,1,1,0), spec$SEX)

data <- data.frame(SEX = c(1,1,1,1,0,0,1,1), STUDY= c(202,100))

head(ys_add_factors(data, spec, SEX, STUDY))

data <- ys_help$data()
spec <- ys_help$spec()

head(ys_add_factors(data, spec))

}
