% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{ys_add_labels}
\alias{ys_add_labels}
\title{Add label attribute to data set columns}
\usage{
ys_add_labels(data, spec, fun = label.ycol, strict = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} to label.}

\item{spec}{yspec object for \code{data}.}

\item{fun}{the function to use for forming \code{label}.}

\item{strict}{if \code{TRUE}, generate an error when \code{data} names do not match
the \code{spec} names; otherwise, only data items in common between \code{data}
and \code{spec} will be labeled.}
}
\description{
This is commonly done prior to saving data to SASxport or other formats
that will track label metadata for a data column.
}
\details{
By default, an error is generated when the column names of the data set
are not identical to the column names in the spec. Set \code{strict = FALSE} to
bypass this error and label only columns that are in common between \code{data}
and \code{spec}.

If the user passes \code{fun} to generate a custom label, the function must take
a single argument, the column \code{ycol} object, and must return the label for
that column as a character vector of length one.
}
\examples{
spec <- ys_help$spec()

data <- ys_help$data()

data <- ys_add_labels(data, spec)

sapply(data, attr, "label")

str(data[,1:5])

}
