% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{ys_add_labels}
\alias{ys_add_labels}
\title{Add label attribute to data set columns}
\usage{
ys_add_labels(data, spec, fun = label.ycol)
}
\arguments{
\item{data}{a \code{data.frame} to label}

\item{spec}{yspec object for \code{data}}

\item{fun}{the function to use for forming \code{label}}
}
\description{
Add label attribute to data set columns
}
\details{
An error is generated if the names of \code{data} are not identical to names
of \code{spec}.

If the user passes \code{fun} to generate a custom label, the function must take
a single argument, the column \code{ycol} object, and must return the label for
that column as a character vector of length one.
}
\examples{
spec <- ys_help$spec()

data <- ys_help$data()

data <- ys_add_labels(data,spec)

sapply(data,attr,"label")

str(data[,1:5])

}
