% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-note.R
\name{ys_col_note}
\alias{ys_col_note}
\title{Create table note text for abbreviated columns}
\usage{
ys_col_note(
  .spec,
  ...,
  .unit = FALSE,
  .title_case = FALSE,
  .sep = ": ",
  .to_string = TRUE,
  .collapse = "; ",
  .width = NULL,
  .type = c("short", "label")
)
}
\arguments{
\item{.spec}{a yspec object.}

\item{...}{passed to \code{\link[=ys_select]{ys_select()}}.}

\item{.unit}{logical; if \code{TRUE}, then append the unit surrounded by parens
to the column definition.}

\item{.title_case}{logical; if \code{TRUE} then the column definition text is
passed through \code{\link[tools:toTitleCase]{tools::toTitleCase()}}.}

\item{.sep}{a string to separate column name and column definition;
usually a space should be included as the terminal character (see default)}

\item{.to_string}{logical; if \code{TRUE}, then a single string is returned}

\item{.collapse}{a string to separate items when \code{.to_string} is \code{TRUE};
usually a space should be included as the terminal character (see default)}

\item{.width}{if \code{numeric} and \code{.to_string} is \code{TRUE}, then the result is
passed through \code{\link[base:strwrap]{base::strwrap()}}.}

\item{.type}{selects if the column definition is generated from calling
\code{\link[=ys_get_short]{ys_get_short()}} or \code{\link[=ys_get_label]{ys_get_label()}}}
}
\value{
A string of length one when \code{.to_string} is \code{TRUE} or a character vector
if \code{.to_string} is \code{FALSE}.
}
\description{
Produces a string with \verb{<col-name>: <col-definition>} format which can be
included in a table note providing a more informative definition with the
column name. The column definition can be generated from the \code{short} name
or the \code{label} (but note that these could frequently be the same).
}
\examples{

spec <- ys_help$spec()

ys_col_note(spec, AST, ALT, SCR, .unit = TRUE) 

}
