% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{ys_document}
\alias{ys_document}
\title{Render a document from one or more specification objects}
\usage{
ys_document(x, type = c("working", "regulatory"), ...)
}
\arguments{
\item{x}{a spec or project object}

\item{type}{the document type}

\item{...}{passed to [render_define] or [render_fda_define]}
}
\description{
Render a document from one or more specification objects
}
