% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{ys_document}
\alias{ys_document}
\title{Render a document from one or more specification objects}
\usage{
ys_document(x, type = c("working", "regulatory"), ...)
}
\arguments{
\item{x}{a spec or project object}

\item{type}{the document type}

\item{...}{passed to \code{\link[=render_define]{render_define()}} or \code{\link[=render_fda_define]{render_fda_define()}}; it is important
to review these help topics to see what other aspects of the document
can be specified; see also \code{details} here.}
}
\description{
This function is a wrapper around \code{\link[=render_define]{render_define()}} and \code{\link[=render_fda_define]{render_fda_define()}}.
}
\details{
Select \code{type} of "working` for a working document that contains more
information in a less-compact layout.  This is a good format for a modeling
and simulation scientist to use day-to-day and for sharing with a sponsor
to review.

Select \code{type} of "regulatory" for a define document that conforms to
submission requirements set by regulatory authorities (e.g. FDA).

Note that \code{x} is usually either a \code{yspec} object or a \code{yproj} object.
You can also pass in the full path to a specification document and
yspec will guess which format it is and render accordingly.

Because \code{...} are passed to \code{\link[=render_define]{render_define()}} and \code{\link[=render_fda_define]{render_fda_define()}},
it is important to review arguments to those functions as well.  Specifically,
please note that the document \strong{title}, \strong{author}, and \strong{date} can
be set, along with the name of the output document, the working document
build directory, and several other aspects of the document can be set
in the call to \link{ys_document}.
}
\section{latex requirements}{


For all document types, the following \code{latex} packages are required:
\enumerate{
\item \code{array}
\item \code{longtable}
\item \code{booktabs}
\item \code{fontenc}
\item \code{mathdesign}
}

Make sure these packages are installed and available when trying to render a document.
}

\examples{

\dontrun{
  ys_document(ys_help$spec())
  ys_document(ys_help$spec(), type = "regulatory")
}
}
\seealso{
\code{\link[=render_define]{render_define()}}, \code{\link[=render_fda_define]{render_fda_define()}}
}
