% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys_filter.R
\name{ys_filter}
\alias{ys_filter}
\title{Select columns from spec object}
\usage{
ys_filter(x, expr, .default = NULL)
}
\arguments{
\item{expr}{an unquoted expression}

\item{.default}{a named list of default look up values}

\item{spec}{a yspec object}
}
\value{
A `yspec` object.
}
\description{
Select columns from spec object
}
\examples{
spec <- ys_help$spec()

ys_filter(spec, is.character(decode))

ys_filter(spec, unit == "kg" | type == "character")

}
