% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-filter.R
\name{ys_filter}
\alias{ys_filter}
\title{Subset spec items using column values}
\usage{
ys_filter(x, expr, .default = NULL)
}
\arguments{
\item{x}{a yspec object}

\item{expr}{an unquoted expression}

\item{.default}{a named list of default look up values}
}
\value{
A `yspec` object.
}
\description{
Subset spec items using column values
}
\details{
The following fields always exist in the spec and are available for 
querying in the filter expression:
- `col`
- `type`
- `discrete`
- `continuous`
- `short`
- `do_lookup`

The following fields will be provided defaults when the filter expression 
is evaluated: 

- `unit` = ""
- `values` = ""
- `decode` = ""
- `covariate` = `FALSE`

In addition to these fields, you can build the filter expression using 
items in the `dots` field.
}
\examples{
spec <- ys_help$spec()

ys_filter(spec, is.character(decode))

ys_filter(spec, unit == "kg" | type == "character")

}
