% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_label}
\alias{ys_get_label}
\title{Get label}
\usage{
ys_get_label(x, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{...}{passed to label methods}
}
\description{
Forms a label for a column (\code{ycol} method) or columns (\code{yspec} method). Use
\link[purrr:map]{purrr::map} or \link[purrr:map]{purrr::map_chr} to simply extract the label field.
}
\examples{
spec <- ys_help$spec()

ys_get_label(spec)

ys_get_label(spec$WT)
\dontrun{
purrr:::map(spec,"label")
}
}
\seealso{
\link{ys_get_short}, \link{ys_get_short}, \link{ys_get_short_unit}
}
