% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_label}
\alias{ys_get_label}
\title{Get label}
\usage{
ys_get_label(x, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{...}{passed to label methods}
}
\description{
Forms a label for a column (`ycol` method) or columns (`yspec` method)
}
\seealso{
[ys_get_short], [ys_get_short], [ys_get_short_unit]
}
