% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{ys_get_lookup}
\alias{ys_get_lookup}
\title{Generate lookup list}
\usage{
ys_get_lookup(x, verbose = FALSE)
}
\arguments{
\item{x}{a yspec object}

\item{verbose}{\code{logical}; print information to the console as the file
is processed}
}
\description{
Generate lookup list
}
\examples{

spec <- load_spec_ex("DEM104101F_PK.yml")
ys_get_lookup(spec)

}
