% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_short}
\alias{ys_get_short}
\title{Get short}
\usage{
ys_get_short(x, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{...}{passed to short methods}
}
\description{
Forms a short for a column (\code{ycol} method) or columns (\code{yspec} method). Use
\link[purrr:map]{purrr::map} or \link[purrr:map]{purrr::map_chr} to simply extract the short field.
}
\examples{
spec <- ys_help$spec()

ys_get_short(spec)

ys_get_short(spec$WT)

\dontrun{
purrr:::map(spec,"short")
}
}
\seealso{
\link{ys_get_unit}, \link{ys_get_label}, \link{ys_get_short_unit}
}
