% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_short}
\alias{ys_get_short}
\title{Get short label}
\usage{
ys_get_short(x, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{...}{passed to label methods}
}
\description{
Get short label
}
\seealso{
[ys_get_unit], [ys_get_label], [ys_get_short_unit]
}
