% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_short_unit}
\alias{ys_get_short_unit}
\title{Get short with unit}
\usage{
ys_get_short_unit(x, .aslist = TRUE, parens = TRUE, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{.aslist}{a named list is returned if \code{TRUE}, otherwise a named
character vector}

\item{parens}{if \code{TRUE}, then parens will be added around any unit with one
or more character}

\item{...}{arguments passed to \code{\link[=ys_get_short]{ys_get_short()}} and \code{\link[=ys_get_unit]{ys_get_unit()}}}
}
\description{
Get short with unit
}
\examples{
spec <- ys_help$spec()

ys_get_short_unit(spec)

ys_get_short_unit(spec$WT)

}
\seealso{
\code{\link[=ys_get_short]{ys_get_short()}}, \code{\link[=ys_get_unit]{ys_get_unit()}}, \code{\link[=ys_get_label]{ys_get_label()}}
}
