% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-get.R
\name{ys_get_short_unit}
\alias{ys_get_short_unit}
\title{Get short with unit}
\usage{
ys_get_short_unit(x, ...)
}
\arguments{
\item{x}{ycol or yspec object}

\item{...}{arguments passed to [ys_get_short] and [ys_get_unit]}
}
\description{
Get short with unit
}
\examples{
spec <- ys_help$spec()

ys_get_short_unit(spec)

ys_get_short_unit(spec$WT)

}
\seealso{
[ys_get_short], [ys_get_unit], [ys_get_label]
}
