% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-gg-labs.R
\name{ys_gg_labs}
\alias{ys_gg_labs}
\title{Automatically label ggplot aesthetics from a yspec object or other source}
\usage{
ys_gg_labs(
  spec = NULL,
  labs = list(),
  x = NULL,
  y = NULL,
  fill = NULL,
  colour = NULL,
  color = NULL,
  col = NULL,
  linetype = NULL,
  lty = NULL,
  shape = NULL,
  ...
)
}
\arguments{
\item{spec}{a yspec object; label data is generated through a call to
\code{\link[=ys_get_short_unit]{ys_get_short_unit()}}.}

\item{labs}{a named list of label data; names correspond to columns
in the data used to make the plot; overrides \code{spec}.}

\item{x}{override label for the x aesthetic; if \code{NULL}, resolved from \code{spec}
or \code{labs}.}

\item{y}{override label for the y aesthetic; if \code{NULL}, resolved from \code{spec}
or \code{labs}.}

\item{fill}{override label for the fill aesthetic; if \code{NULL}, resolved from
\code{spec} or \code{labs}.}

\item{colour}{override label for the colour aesthetic; if \code{NULL}, resolved
from \code{spec} or \code{labs}; \code{color} and \code{col} are accepted as aliases.}

\item{color}{alias for \code{colour}.}

\item{col}{alias for \code{colour}.}

\item{linetype}{override label for the linetype aesthetic; if \code{NULL},
resolved from \code{spec} or \code{labs}; \code{lty} is accepted as an alias.}

\item{lty}{alias for \code{linetype}.}

\item{shape}{override label for the shape aesthetic; if \code{NULL}, resolved from
\code{spec} or \code{labs}.}

\item{...}{additional arguments passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}.}
}
\value{
A gg object that can be added to a ggplot with \code{+}.
}
\description{
Automatically label ggplot aesthetics from a yspec object or other source
}
\examples{
if(requireNamespace("ggplot2")) {
  library(ggplot2)
  
  spec <- ys_help$spec()
  
  data <- ys_help$data()
  
  p <- ggplot(data, aes(TIME, DV)) + geom_point()
  
  p + ys_gg_labs(spec)
}

}
