% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-gg-labs.R
\name{ys_gg_labs}
\alias{ys_gg_labs}
\title{Automatically label ggplot axes from a yspec object or other source}
\usage{
ys_gg_labs(
  spec = NULL,
  labs = list(),
  xcol = NULL,
  ycol = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{spec}{a yspec object; axis data is generated through a call to 
[ys_get_short_unit()].}

\item{labs}{a named list of axis title data; names correspond to columns 
in the data used to make the plot; overrides `spec`.}

\item{xcol}{the name of the x-axis data column; overrides `labs` and `spec`.}

\item{ycol}{the name of the y-axis data column; overrides `labs` and `spec`.}

\item{xlab}{the complete title to use for the x-axis; overrides `xcol` and 
`spec`.}

\item{ylab}{the complete title to use for the y-axis; overrides `ycol` and 
`spec`.}

\item{...}{additional arguments passed to [ggplot2::labs()].}
}
\value{
A gg object that can be added to a ggplot with `+`.
}
\description{
Automatically label ggplot axes from a yspec object or other source
}
