% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-gg-labs.R
\name{ys_gg_labs}
\alias{ys_gg_labs}
\title{Automatically label ggplot axes from a yspec object or other source}
\usage{
ys_gg_labs(
  spec = NULL,
  labs = list(),
  x = NULL,
  y = NULL,
  fill = NULL,
  col = NULL,
  lty = NULL,
  shape = NULL,
  ...
)
}
\arguments{
\item{spec}{a yspec object; axis data is generated through a call to 
[ys_get_short_unit()].}

\item{labs}{a named list of axis title data; names correspond to columns 
in the data used to make the plot; overrides `spec`.}

\item{x}{passed to [ggplot2::labs()] if character.}

\item{y}{passed to [ggplot2::labs()] if character.}

\item{fill}{passed to [ggplot2::labs()] if character.}

\item{col}{passed to [ggplot2::labs()] if character.}

\item{lty}{passed to [ggplot2::labs()] if character.}

\item{shape}{passed to [ggplot2::labs()] if character.}

\item{...}{additional arguments passed to [ggplot2::labs()].}
}
\value{
A gg object that can be added to a ggplot with `+`.
}
\description{
Automatically label ggplot axes from a yspec object or other source
}
