% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-gg-labs.R
\name{ys_gg_labs}
\alias{ys_gg_labs}
\title{Automatically label ggplot aesthetics from a yspec object or other source}
\usage{
ys_gg_labs(
  spec = NULL,
  labs = list(),
  x = NULL,
  y = NULL,
  fill = NULL,
  colour = NULL,
  color = NULL,
  col = NULL,
  linetype = NULL,
  lty = NULL,
  shape = NULL,
  ...
)
}
\arguments{
\item{spec}{a yspec object; label data is generated through a call to
[ys_get_short_unit()].}

\item{labs}{a named list of label data; names correspond to columns
in the data used to make the plot; overrides `spec`.}

\item{x}{override label for the x aesthetic; if `NULL`, resolved from `spec`
or `labs`.}

\item{y}{override label for the y aesthetic; if `NULL`, resolved from `spec`
or `labs`.}

\item{fill}{override label for the fill aesthetic; if `NULL`, resolved from
`spec` or `labs`.}

\item{colour}{override label for the colour aesthetic; if `NULL`, resolved
from `spec` or `labs`. `color` and `col` are accepted as aliases.}

\item{color}{alias for `colour`.}

\item{col}{alias for `colour`.}

\item{linetype}{override label for the linetype aesthetic; if `NULL`,
resolved from `spec` or `labs`. `lty` is accepted as an alias.}

\item{lty}{alias for `linetype`.}

\item{shape}{override label for the shape aesthetic; if `NULL`, resolved from
`spec` or `labs`.}

\item{...}{additional arguments passed to [ggplot2::labs()].}
}
\value{
A gg object that can be added to a ggplot with `+`.
}
\description{
Automatically label ggplot aesthetics from a yspec object or other source
}
