% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{data}
\name{ys_help}
\alias{ys_help}
\title{yspec help object}
\format{An object of class \code{environment} of length 11.}
\usage{
ys_help
}
\description{
\code{ys_help} is an environment containing functions that can be called to
provide information and examples for how yspec works.  \code{specification()},
\code{internal_db()} and \code{analysis1()} will open up markdown or yaml files
using \link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile} if that package is installed and
\link{file.show} in case it is not.
}
\details{
\itemize{
\item \code{file()} the file name and path of an example yspec file
\item \code{spec()} generates a yspec object from \code{file()}
\item \code{data()} is a function that generates a data set that can be checked
against \code{spec()}
\item \code{csv()} is the file name (and path) to the data set generated by \code{data()}
\item \code{proj()} is a function that generates a project object that can be
rendered into a document
\item \code{specification()} is a markdown document explaining what fields to include
in the yaml specification
\item \code{internal_db()} is the internal column lookup data base
\item \code{analysis1()} is an example data specification file
}
}
\examples{
spec <- ys_help$spec()

file <- ys_help$file()

data <- ys_help$data()  

csv  <- ys_help$csv()

\dontrun{
proj <- ys_help$proj()

specif <- ys_help$specification()

inter <- ys_help$internal_db()

analysis1 <- ys_help$analysis1()
}

}
\keyword{datasets}
