% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-filter.R
\name{ys_join}
\alias{ys_join}
\title{Join yspec objects together}
\usage{
ys_join(left, right, ...)
}
\arguments{
\item{left}{a yspec object}

\item{right}{a yspec object}

\item{...}{more yspect objects}
}
\value{
A single yspec object
}
\description{
Join yspec objects together
}
\details{
All inputs must be `yspec` objects. When the `right` spec (or specs passed
under `...`) is joined to the `left` spec, columns that exist in both 
`left` and `right` are removed from `right` before joining.
}
\examples{

spec <- ys_help$spec()

l <- ys_select(spec, WT, BMI)
r <- ys_select(spec, TIME, TAD)
rr <- ys_select(spec, EVID, MDV, CMT, BMI)

ys_join(l, r)

ys_join(l, r, rr)

}
\seealso{
[ys_rename()], [ys_filter()], [ys_select()]
}
