% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spec.R
\name{ys_load}
\alias{ys_load}
\alias{ys_load_file}
\alias{load_spec}
\title{Load a data specification file}
\usage{
ys_load(file, ...)

ys_load_file(file, data_path = NULL, data_stem = NULL, ...)

load_spec(...)
}
\arguments{
\item{file}{name of yaml file containing specification}

\item{...}{other arguments to update `SETUP__`}

\item{data_path}{optional path to data sets}

\item{data_stem}{optional alternate stem for data files}
}
\description{
Load a data specification file
}
