% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ys_sanitize}
\alias{ys_sanitize}
\title{Default function to sanitize text for TeX documents}
\usage{
ys_sanitize(x)
}
\arguments{
\item{x}{text to sanitize}
}
\description{
Based on \link[xtable:sanitize]{xtable::sanitize}.
}
\examples{
ys_sanitize("Concentration ($\\\\mu$g)")
ys_sanitize("H&M")

}
