% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-yspec.R
\name{ys_select}
\alias{ys_select}
\title{Select a subset of columns from a yspec object}
\usage{
ys_select(x, ...)
}
\arguments{
\item{x}{a yspect object}

\item{...}{unquoted columns to select}
}
\description{
Select a subset of columns from a yspec object
}
