% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ys-tidy.R
\name{ys_select}
\alias{ys_select}
\title{Select a subset of columns from a yspec object}
\usage{
ys_select(.x, ...)
}
\arguments{
\item{.x}{a yspec object}

\item{...}{unquoted columns to select}
}
\value{
A \code{yspec} object
}
\description{
Select a subset of columns from a yspec object
}
\examples{

spec <- ys_help$spec()

ys_select(spec, WT, AGE, ALB)

ys_select(spec, Wt = WT, AGE)

}
