% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{ys_table}
\alias{ys_table}
\title{Create a table from yspec object}
\usage{
ys_table(
  spec,
  fun = NULL,
  tex = TRUE,
  ns = c("tex", "define"),
  widths_ = c(0.75, 1.95, 0.6, 2.15),
  ...
)
}
\arguments{
\item{spec}{a \code{yspec} object.}

\item{fun}{a function to format a TeX table; if \code{NULL} (the default), the
table will be rendered using \code{\link[=fda_table]{fda_table()}}.}

\item{tex}{logical; if \code{TRUE}, switch to \code{tex} namespace; see \code{ns} argument.}

\item{ns}{character vector of namespaces to load before generating tables.}

\item{widths_}{passed to \code{\link[=fda_table]{fda_table()}} when \code{fun} is \code{NULL}; these are
slightly modified from the \code{\link[=fda_table]{fda_table()}} default (see \code{examples}); note
the trailing underscore in the argument name; these shouldn't need to be
changed for most use.}

\item{...}{additional arguments passed to \code{fun}.}
}
\value{
The table text generated from \code{fun}.
}
\description{
The primary use case for this function is for creating TeX tables which can
be included in a report Appendix. See more in \code{details}.
}
\details{
By default, the table code is rendered using \code{\link[=fda_table]{fda_table()}}; this should be
used in most cases. \code{\link[=fda_table]{fda_table()}} returns the table in the \code{longtable}
environment. This can be included in a report with \verb{\\input\{<file.tex>\}}.

At this time, there is no mechanism for generating a caption for tables
generated using \code{\link[=fda_table]{fda_table()}}; the intended use is to include the table in
an appendix, with caption information given in plain text in the appendix.
}
\examples{
spec <- ys_help$spec()
tab <- ys_table(spec)
writeLines(text = tab, con = tempfile(fileext=".tex"))

formals(fda_table)$widths
formals(ys_table)$widths_

}
