% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_define.R
\name{ys_table}
\alias{ys_table}
\title{Create a table from yspec object}
\usage{
ys_table(spec, fun = NULL, ...)
}
\arguments{
\item{spec}{a \code{yspec} object}

\item{fun}{a function to format a TeX table; if \code{NULL} (the default), the
table will be rendered using \code{\link[=fda_table]{fda_table()}}}

\item{...}{additional arguments passed to \code{fun}}
}
\value{
The table text generated from \code{fun}; if \code{file} was given to save the table,
then the table text is returned invisibly.
}
\description{
The primary use case for this function is for creating TeX tables which can
be included in a report Appendix. See more in \code{details}.
}
\details{
By default, the table code is rendered using \code{\link[=fda_table]{fda_table()}}; this should be
used in most cases. \code{\link[=fda_table]{fda_table()}} returns the table in the \code{longtable}
environment. This can be included in a report with \verb{\\input\{<file.tex>\}}.

At this time, there is no mechanism for generating a caption for tables
generated using \code{\link[=fda_table]{fda_table()}}.
}
