% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aaaa.R
\docType{package}
\name{yspec}
\alias{yspec}
\alias{yspec-package}
\title{yspec: data specification documents from yaml.}
\description{
yspec: data specification documents from yaml.
}
\section{Options}{

\itemize{
\item \code{ys.sanitize} a function to use for sanitizing text before processing wiht
latex; see \code{\link[yspec:pander_table]{yspec::pander_table()}}
\item \code{ys.col.length} the maximum number of characters in a data frame column
name; this defaults to 8 so that columns with 9 or more characters will
generate an error on load.
\item \code{ys.fct.suffix} the suffix to add to a column name, used by
\code{\link[yspec:yspec_add_factors]{yspec::yspec_add_factors()}}
}
}

