% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aaaa.R
\docType{package}
\name{yspec}
\alias{yspec-package}
\alias{yspec}
\title{yspec: data specification documents from yaml}
\description{
After writing data specification details in yaml format, read in as an
object in R, query data set details, check / validate data sets, and
write out data definition documents.  To get started, check out the
\href{../doc/index.html}{vignettes}.
}
\section{Workflow}{

\itemize{
\item Write your document in yaml format
\itemize{
\item Include a \verb{SETUP__:} block at the top to include meta information
\item Use \code{lookup}s from either the internal library or your own lookup file
\item Syntax reference \href{../doc/reference.html}{here}
\item Use \link{ys_help} to access examples
}
\item Use \code{\link[=ys_load]{ys_load()}} to read in the yaml file
\itemize{
\item Use the \code{.verbose} to see more information while this is in progress
}
\item Once the \code{spec} object is loaded
\itemize{
\item \code{spec$col} to see the definition of \code{col}
\item \code{summary(spec)} to see a summary
\item \code{\link[=yspec_add_factors]{yspec_add_factors()}} to create factors in the data set
}
\item Use \code{\link[=ys_check]{ys_check()}} to check a data frame against the spec
\item Use \code{\link[=ys_document]{ys_document()}} to render the specification object as a pdf file
\item Use \code{\link[=ys_project]{ys_project()}} to create a collection of individual data specification
objects; this also can be rendered as a pdf file with \code{\link[=ys_document]{ys_document()}}
\item Get help: Use \link{ys_help} to get further help and documentation in your R
session
}
}

\section{Package-wide options}{

\itemize{
\item \code{ys.sanitize} a function to use for sanitizing text before processing with
latex; see \code{\link[=pander_table]{pander_table()}}
\item \code{ys.fct.suffix} the suffix to add to a column name, used by
\code{\link[=yspec_add_factors]{yspec_add_factors()}}
\item \code{ys.require.label} if \code{TRUE}, an error will be generated whenever a column
is specified without a label
}

\strong{NOTE}: \code{ys.col.len} was an available in previous versions to set the
maximum number of characters allowable in a column name. This options has
been deprecated.  Please use \code{max_char_col} in \verb{SETUP__:} instead.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://metrumresearchgroup.github.io/yspec}
  \item \url{https://github.com/metrumresearchgroup/yspec}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/yspec/issues}
}

}
\author{
\strong{Maintainer}: Kyle T Baron \email{kyleb@metrumrg.com} (\href{https://orcid.org/0000-0001-7252-5656}{ORCID})

Other contributors:
\itemize{
  \item Katherine Kay \email{katherinek@metrumrg.com} [contributor]
  \item Eric Anderson \email{andersone@metrumrg.com} [contributor]
  \item Sean Wilson \email{copernican@gmail.com} [contributor]
  \item Samuel Callisto \email{samc@metrumrg.com} [contributor]
  \item Metrum Research Group [copyright holder]
}

}
