% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_factor.R
\name{yspec_add_factors}
\alias{yspec_add_factors}
\alias{yspec_make_factor}
\title{Add factors to data set based on spec information}
\usage{
yspec_add_factors(.data, .spec, ..., .all = TRUE,
  .suffix = getOption("ys.fct.suffix", "_f"))

yspec_make_factor(values, x)
}
\arguments{
\item{.data}{the data set to modify}

\item{.spec}{a yspec object}

\item{...}{unquoted column names for modification}

\item{.suffix}{used to make the column name for the factors}

\item{values}{a vector of values to convert to a factor}

\item{x}{a ycol object}
}
\description{
Add factors to data set based on spec information
}
\details{
Note that `.suffix` can be chosen using option `ys.fct.suffix`.
}
\examples{

spec <- load_spec_ex()

yspec_make_factor(c(1,0,1,1,1,0), spec$SEX)

data <- data.frame(SEX = c(1,1,1,1,0,0,1,1), STUDY= c(202,100))

yspec_add_factors(data, spec, SEX, STUDY)

}
